Move special keys with fn closer to index finger
The idea is to use the same position of the keys as the MacBook keyboard. 
Considering that one way to reach those keys is moving the hand to the 
corner of the keyboard I am keeping the same disposition but keeping 
the hand on the home keys using the fn key.